requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
groq==0.9.0
python-dotenv==1.0.1
pandas==2.2.2
